# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: YDB Installer API
  version: 0.1.0
  description: |
    Public REST API for installation session lifecycle (FR-API-001–003).
    Machine-readable contract placeholder; expand to match handlers in api/.
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: localhost:8443
paths:
...
  /sessions:
    get:
      summary: List installation sessions
      operationId: listSessions
      parameters:
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
      responses:
        '200':
          description: OK
    post:
      summary: Create session
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode: { type: string, enum: [interactive, batch] }
                title: { type: string }
      responses:
        '201':
          description: Created
  /sessions/{sessionID}:
    parameters:
      - $ref: '#/components/parameters/sessionID'
    get:
      summary: Get session
      operationId: getSession
      responses:
        '200':
          description: OK
    patch:
      summary: Update draft metadata
      operationId: patchSession
      responses:
        '204':
          description: No Content
    delete:
      summary: Delete session
      operationId: deleteSession
      responses:
        '204':
          description: No Content
...
components:
  parameters:
    sessionID:
      name: sessionID
      in: path
      required: true
      schema:
        type: string
        format: uuid